% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichGopck.R
\name{automatic_GO_enrich}
\alias{automatic_GO_enrich}
\title{automatic_GO_enrich
The main function of the package that will iterate the list of file names to run the GO analysis}
\usage{
automatic_GO_enrich(
  x = dir(),
  spcode,
  keytype = "ENSEMBL",
  orderby = "",
  dotplotgenes = 30,
  fontSize = 8,
  genekeyPos = 0,
  maxnumberofgenes = 5000,
  GO_pvalue = 0.05,
  KEGG = TRUE,
  GOALL = TRUE,
  GOBP = TRUE,
  GOMF = TRUE,
  GOCC = TRUE,
  writeTable = TRUE,
  writeplot = TRUE
)
}
\arguments{
\item{x}{the data.frame or vector that will contain the name/path of files of differentially expressed genes}

\item{spcode}{a char of the species code of the gene entries. Call listspecies() to look for the species code available}

\item{keytype}{a char for the key type of genes. The possible values are the following: ENSEMBL, ENTREZID, REFSEQ, SYMBOL, ENSEMBLTRANS, ENSEMBLPROT. The default value is ENSEMBL.}

\item{orderby}{char containin name of column for which the gene keys must be ordered. If left empty, the order of the input list will be used.}

\item{dotplotgenes}{integer containing the number of genes to the dotplot, default = 30}

\item{fontSize}{integer containing the point of font of the plot, default = 8}

\item{genekeyPos}{integer containing the number of column of the gene identifiers. If no number set, we will try to find it accordingly to the keytype.}

\item{maxnumberofgenes}{integer containing the maximum number of genes to be evaluated by the GO algorithms. The default value is 5000.}

\item{GO_pvalue}{number containg the threshold for the GO enrichment algorithms. The default value is 0.05.}

\item{KEGG, }{GOALL, GOBP, GOMF, GOCC, writeTable and writeplot are logical values. They determine wheter enrichment will be runned and if the dotplot and table will be writed. The default of all values is TRUE.}
}
\description{
automatic_GO_enrich
The main function of the package that will iterate the list of file names to run the GO analysis
}
\author{
Soares, Luis
}
