% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_database.R
\name{choose_database}
\alias{choose_database}
\title{choose_database}
\usage{
choose_database(db_search = NULL)
}
\arguments{
\item{db_search}{(Default = NULL), is the string pattern to be matched against the list of enrichR databases. Any matching DBs will be returned.}
}
\value{
List of database names as a character vector.
}
\description{
It allows the user to choose the databases on which to perform the enrichment analysis. Either it returns all the possible databases or a subset of them.
}
\examples{
\dontrun{
choose_database(db_search = "KEGG")
}
}
