% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparisons}
\alias{comparisons}
\title{Example comparison data.}
\format{
A data frame with 6 rows and 2 variable:
}
\usage{
comparisons
}
\description{
It is the comparison table needed by DESeq2, constituted by the
column \emph{treatment} and by the column \emph{control}.
}
\keyword{datasets}
