% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Example raw count data.}
\format{
A data frame with 29744 rows and 19 variable:
}
\usage{
counts
}
\description{
Genes as rows and samples as columns.
}
\keyword{datasets}
