% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{Example group data.}
\format{
A data frame with 18 rows and 2 variable:
}
\usage{
groups
}
\description{
It is the sample information table needed by DESeq2, constituted by the
column \emph{sample}, including sample barcodes and by the column \emph{group}
including the condition to which each sample belongs.
}
\keyword{datasets}
