% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichGopck.R
\docType{data}
\name{internenv}
\alias{internenv}
\title{geneIDAdvise
Intern function to format the print output}
\format{
An object of class \code{environment} of length 1.
}
\usage{
internenv
}
\description{
geneIDAdvise
Intern function to format the print output
}
\keyword{datasets}
