% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliGO.R
\name{lolliGO}
\alias{lolliGO}
\title{lolliGO}
\usage{
lolliGO(
  enrich_tables,
  title = NULL,
  outfolder = NULL,
  outfile = "lolliGO.png",
  from_autoGO = TRUE
)
}
\arguments{
\item{enrich_tables}{Dataframe containing the enrichment results or a path to your .tsv file containing the enrichment results. Columns 'Term' and 'Adjusted.P.Value' are required.}

\item{title}{Default to NULL, only specify if from_autoGO is FALSE. When enrich_tables is not from autoGO and thus from read_enrich_tables, the user can specify title and subtitle of the plot as 2-element character vector, for example: c("This is the title", "this is the subtitle")}

\item{outfolder}{Default to NULL, only specify if from_autoGO is FALSE. The name to assign to the folder for output saving.}

\item{outfile}{Default to "lolliGO.png", is ignored if from_autoGO is TRUE. The name of the lolli filename.}

\item{from_autoGO}{Default is TRUE, set to FALSE if the enrichment tables you want to use are not from a differential expression analysis.}
}
\value{
No return value. Files will be produced as part of normal execution.
}
\description{
The function lolliGO.R implement the lollipop plot of the first 20 enrichment terms.

For each enrichment result table a lollipop plot is produced. Results are stored in the "enrichment_plots" subfolder for each comparison.
}
\examples{
\dontrun{
lolliGO(
  enrich_tables = enrich_tables,
  title = NULL,
  outfolder = NULL,
  outfile = NULL,
  from_autoGO = TRUE
)
}
}
